ADFFACES-448 portal+skinning: suppress stylesheet if requested skin-id+skin's stylesheetid are found
If a portal is running Trinidad it may want its portlets running Trinidad to use the same skin as it is running instead of the default portal skin. To tell the portlet what skin to use, the portal sends some information via the RequestMap. The information is the skin-id and the skin's stylesheetdocument's id.
We will check to see if we have this skin and if its stylesheet document matches exactly and we will use this skin. Otherwise we will use the default portal skin.

The portal may want to tell the portlet to not render out the css url (see StyleSheetRenderer) since it will render out its own url to help performance. We will check another RequestMap parameter for this.

I propose adding this code and will check for these RequestMap parameters:

oracle.apache.myfaces.trinidad.skin.suppressStylesheet
oracle.apache.myfaces.trinidad.skin.id
oracle.apache.myfaces.trinidad.skin.stylesheet.id

Added this Skin API:
  /**
   * Returns the id of the Skin's stylesheet document.
   */
  abstract public String getStyleSheetDocumentId(RenderingContext arc);