https://issues.apache.org/jira/browse/TRINIDAD-628
purple skin's navigationPane tabs does not render correctly

The jspx code in panelPageSkinDemo.jspx with its vertical panelLayout was screwing up the tabs.
I also added skin demos of the navigationPane tabs component in purple and beach. The beach tabs
use images that are blue/gold like the old 'oracle' skin. This will help write the documentation of the tabs
in skin-selectors.