TRINIDAD-700 - EL isn't evalutaed for the datestyle / timestyle attrs of tr:convertDateTime
fix is inside the constructor of the Converter itself:

Since we're explicitly setting the time style
and date style and type, the value binding will be ignored.
Instead of using setters in the constructor, we should put
that default into the getters. 

True for the Trinidad API NumberConverter clazz as well.