TRINIDAD-749 Launching modal dialog when the launcher window body has position:absolute in IE makes launcher page look corrupt
Removed the code where in IE and launching a modal dialog, the opacity of the body element was being changed.  Instead, created a new DIV that appears on top of the page content that is white and 50% opaque to achieve the same visual result.

Now, whenever there are other semi-opaque elements on the page and if the body uses absolute positioning, those semi-opaque elements will be drawn correctly--no incorrectly masked areas and thus a non-corrupt appearance.

Also added some svn:ignore property changes for JDeveloper project files so they won't accidentally get committed.