Issue 906 optimize getClientId - use per thread shared StringBuilder
In getClientId/getContainerClientId we create a new stringBuilder.

However when creating a client id we only use one stringBuilder at a time, so have a single instance of a StringBuilder saved on thread local.

Have a utility to get this instance, and any time a user gets the instance, set the length of the stringBuilder to 0.