TRINIDAD-908 optimize invokeOnComponent in a namingContainer
Avoid processing children based on id's in invokeOnComponent.

In UIXCollection it implements invokeOnComponent, and unless the client id passed in starts with the client id of the collection, it doesn't process the children. However not all our namingContainers are doing this.

NamingContainers that don't avoid processing children based on id's in invokeOnComponent

    * UIXComponentRef
    * UIXDecorateCollection
    * UIXMenu
    * UIXSubform