TRINIDAD-833 Ability to skin branding facet for tr:panelPage
Added af|panelPage::branding which applies to the wrapper element around the content placed in the "branding" facet of the tr:panelPage component.  In the simple skin, this has no default styles.  In the minimal skin, this provides a 4px padding style by default.

Replaced a few more corporate logos with generic logos.