TRINIDAD-927 panelBox with an icon or title does not support stretching to 100% width.
The panelBox without an icon and title will support stretching of its contents to be 100% wide when you have an inlineStyle="width:100%;".  However, if you have either an icon or title text present, there is a table element that wraps the header and content and this table didn't have its width set to 100%.  Now, both scenarios will support being 100% wide.