TRINIDAD-1126 - Contention at java.util.regex.Pattern.matcher
Problem is, String.replace(CharSeq, CharSeq) uses internally Pattern.matcher, which has a (serious) bug, b/c it does synchronized(this), which causes PERF issues.

JDK bug is known, already:
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6436458

Not sure if it is fixed in an update of Java5 or 6, but it is fixed in Java7.
For what it's worth, Apache Harmony, doesn't have that problem.

Fix:
-Introduced a StringUtils.replace(String, String, String), taken from Commons LANG.
-used this static util, instead of "nativ" java language feature(the String.replace(CharSq, CharSq)).
-In StyleUtils we got rid of Pattern.matcher().replaceAll. This is exactly, what the String.replace() in question does. Using the "new", borrowed, utility here as well