TRINIDAD-1155 - Tree initiallyExpanded="true" attribute leaves a dangling currency
If initiallyExpanded="true" attribute is set on the tree, it leaves a dangling currency pointing to the first row. This causes unexpected results since the clientId of the tree now contains "0" (currency of the first row). The fix is to make sure that after the first rowKey is expanded in the tree, we restore the currency to the initial value.

Thanks to Venkata Guddanti for the patch.