TRINIDAD-1224 - Support converters for inputFile componet
The "problem" is, that getAsObject(....) only takes String, instead of Object.
Now, the SimpleInputFileRenderer uses a convention, to pass in a (unique) identifier. Before that, the actual file is stored in the requestMap:
 =>externalContext.getRequestMap().put(someKeyForTheUploadedFile, theActualUploadedFileImplementation);

So... a custom converter would look for the key in the reqMap to do whatever "transformation" is needed.