TRINIDAD-1360 error in doubleconverter
Doubleconverter ignores locale, so it just does this on the server

* getAsString() --> return Double.toString(((Number)value).doubleValue());
* getAsObject() --> return Double.valueOf(stringValue);

On the client it does something similar to the server for getAsString.

However on the client for getAsObject it's calling _decimalParse, which uses localeSymbols to parse, which means it's using locale specific formatting on the client to interpret the string to an object. 

Changed the _decimalParse function to take an optional boolean to ignore the locale symbols for parsing and send in true for double and float converter.