TRINIDAD-1635 Trinidad 2 - nullpointerexception thrown when partial state saving turned on
Support using the token cache when there is a stateManagementStrategy available

Because StateManagementStrategy.restoreView does not support passing in state, we are pushing cached state on the request and having the CoreResponseStateManager return that if it's available. I will log a JSF bug asking that StateManagementStrategy.restoreView take the state object returned from StateManagementStrategy.saveView.