TRINIDAD-1547  @agent in skinning should allow for a range of agent versions also, TrinidadAgent.getApplication() changed from an int to an enum, and non-TrinidadImpl users of this api (ADF Faces) will need to change to using the enum commit for Blake Sullivan
Also fix bug where the css file was missing most of the selectors from the .xss skinning files.
The reason was that the AgentAtRuleMatcher was being checked, even for StyleSheetNode's that did not have browsers specified.
Therefore only the StyleSheetNodes with browsers specified and that matched the user's browser were being included in the css.