TRINIDAD-885  - Provide selector to show navigation bar on bottom of table
The old ADF Faces table had a skin property like:
af|table{ -ora-repeat-control-bar: true; }

but the Apache MyFaces Trinidad table did not expose an equivalent:
af|table { -tr-repeat-control-bar: true; }

While there was an existing variable for this skin property and some TableRenderer code to decide which bar to render, it wasn't actually functional or documented.

I've added documentation for the skin property and have enabled the TableRenderer to use this property to decide whether to repeat the control bar.
The default remains to not repeat the control bar and you can enable it by adding this rule in your skin:
af|table { -tr-repeat-control-bar: true; }