TRINIDAD-636 Skinning an icon does not pick up base skin's non-overridden properties
Do not create the Icon Object in SkinStyleSheetParserUtils.java, like we did before.
Add StyleNode object to IconNode, so that an IconNode can have Styles.
Then use the reconciling code that we have in place in StyleSheetDocument for for StyleNodes to also work for IconNodes.
Test this out in purple skin.