TRINIDAD-1851 - Test failures when building against Mojarra 2.0.3
looks like a valid call against getValueExpression() has been added to their NumberConverter. The issue with our mock approach is that we explicitly say ALLOWED or NOT ALLOWED.
I added some code to indicate that we do not care about getValueExpression() being called or not. This allows us building against different MyFaces and Mojarra versions