TRINIDAD-1893 trh:tablelayout not a layout table in screen reader mode
trh:tableLayout should be identified as a layout table (having role="presentation") when it is rendered in screen reader mode.

TableLayoutRenderer currently calls the OutputUtils.renderDataTableAttributes method to render its table, and does not get this role assignment. The TableLayoutRenderer should be updated to call the OutputUtils.renderLayoutTableAttributes method instead, so it will get the role assignment.

I'll create and upload a patch.

Thanks to Dave Robinson for the patch.