TRINIDAD-1895 Issue with UIXIterator visitData implementation
The UIXIterator 'visitData' method is called during 'visitTree' invocation to visit the stamped children of UIXIterator.

The implementation uses an "IndexedRunner"/"KeyedRunner" to loop over all or some rows and visit the contents of the row. The issue is that the code skips over direct stamped children of the UIXIterator and instead visits the grandchildren.

Looks like the code is assuming that direct children of UIXIterator are unstamped columns (UIXColumn) and should be skipped over during stamping-visit. This assumption is correct for sub-classes of UIXIterator but not for UIXIterator. UIXIterator does not have columns.

The visiting logic is already using a NoColumnFacetsVisitContext. So on invokeVisitCallback, it skips over any columns and visits the column children. It also handles nested columns.

Thanks Kamran Kashanian for the patch.