TRINIDAD-2234 Pregeneration of skin style sheets TRINIDAD-2235 Skinning: stable names for generated style sheets
Pregeneration of skin style sheet is now supported via the new skin pregeneration "service".  This is implemented as internal view ("/-tr-pregenerate-skins") which triggers pregeneration of the physical style sheets (.css files) corresponding to a skin that is specified via a request parameter.

See the "Pre-generating Skin Style Sheets" section of the Skinning chapter in the Trinidad Developer's Guide for more details.

In addition, the file names for generated style sheets now use a "stable" naming structure that:

1. Is guranteed to be stable/consistent across runs.  And...
2. Is well-defined, and thus targetable by regular expressions.

See the "Style Sheet File Names" section of the Skinning chapter in the Trinidad Developer's Guide for more details.