TRINIDAD-2340 LocaleElementsResourceLoader init dependency on request path
Avoid calling CoreRenderKitResourceLoader.getLocaleElementsURI() during resource loader initialization, as it is not safe to depend on the request uri at this time.  Fortunately, this call was entirely unnecessary, so we can simply remove it with no negative impact.

Added a small unit test that verifies that the CoreRenderKitResourceLoader (and thus the LocaleElementsResourceLoader) can be instantiated without access to a FacesContext (and thus is not dependent on any request-specific state).